+++ package.json
@@ -41,6 +41,10 @@
     "stylelint-scss": "^3.21.0"
   },
   "devDependencies": {
+    "babel-eslint": "^10.1.0",
+    "eslint": "^7.32.0",
+    "eslint-config-airbnb-base": "^14.2.1",
+    "eslint-plugin-import": "^2.25.4",
     "hint": "^6.1.9"
   },
   "scripts": {
